Change offset in memtable_entry to unsigned int
The offset field in memtable_entry has been changed from an
int to an unsigned int because offset is always a positive number.
This forced a change to memtable_read, it now accepts the address
of where to store the offset and returns a positive number if found
or -1 if not.