Add tombstone parameter to segf_append
A tombstone is metadata assoicated with a key value pair. At this time
all it is used for is to determine if the kv pair is deleted. So if
tombstone == 1 then the kv pair is an existing kv pair that is now being
deleted from the segment file. If tombstone == 0 then the kv pair is a new
kv pair. The repopulate, merge, and compact functions will ignore kv pairs
with a tombstone of 1.