Change order of operations in segf_append
Old version of segf_append would write the key value pair to the segment
file then write key offset pair to the memtable. If there was an error adding
to the memtable (no memory for example) the key value pair would not be
removed from the segment file thus making the segment file and associated
memtable inconsistent with each other.

New version of the segf_append adds the key offset pair to the memtable first
then writes the key value pair to the segment file. If there is an error doing
the write then the key offset pair is removed from the memtable. As a result,
an error adding a new key value pair to the DB does not make the segment file
and its memtable inconsistent with each other.